test: split humongous test file
