+++ src/fat/fs.rs
@@ -2486,3 +2486,234 @@ where
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use akin::akin;
+    use test_log::test;
+
+    pub static MINFS: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/tests/common/imgs/minfs.img"
+    ));
+    pub static FAT12: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/tests/common/imgs/fat12.img"
+    ));
+    pub static FAT16: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/tests/common/imgs/fat16.img"
+    ));
+    pub static FAT32: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/tests/common/imgs/fat32.img"
+    ));
+
+    pub static BEE_MOVIE_SCRIPT: &str = include_str!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/tests/common/bee movie script.txt"
+    ));
+
+    /// Check the reserved FAT entries (first two) for expected values
+    #[test]
+    #[expect(non_snake_case)]
+    fn check_FAT_offset() {
+        use crate::fat::BootRecord;
+
+        use std::io::Cursor;
+
+        let mut storage = FromStd::new(Cursor::new(FAT16.to_owned())).unwrap();
+        let fs = FileSystem::new(&mut storage, FSOptions::new()).unwrap();
+
+        let fat_offset = match &*fs.boot_record.borrow() {
+            BootRecord::Fat(boot_record_fat) => boot_record_fat.first_fat_sector(),
+            BootRecord::ExFAT(_boot_record_exfat) => unreachable!(),
+        };
+
+        // we manually read the first and second entry of the FAT table
+        fs.load_nth_sector(fat_offset.into()).unwrap();
+
+        let first_entry = u16::from_le_bytes(fs.sector_buffer.borrow()[..2].try_into().unwrap());
+        let media_type = if let BootRecord::Fat(boot_record_fat) = &*fs.boot_record.borrow() {
+            boot_record_fat.bpb._media_type
+        } else {
+            unreachable!("this should be a FAT16 filesystem")
+        };
+        assert_eq!(u16::MAX << 8 | u16::from(media_type), first_entry);
+
+        // FIXME: this may not necessarily be full of 1s
+        let second_entry = u16::from_le_bytes(fs.sector_buffer.borrow()[2..4].try_into().unwrap());
+        assert_eq!(u16::MAX, second_entry);
+    }
+
+    /// Ensure that changes are mirrored to both FAT tables
+    #[test]
+    #[expect(non_snake_case)]
+    fn FAT_tables_after_write_are_identical() {
+        use std::io::Cursor;
+
+        let mut storage = FromStd::new(Cursor::new(FAT16.to_owned())).unwrap();
+        let fs = FileSystem::new(&mut storage, FSOptions::new()).unwrap();
+
+        assert!(
+            fs.FAT_tables_are_identical().unwrap(),
+            concat!(
+                "this should pass. ",
+                "if it doesn't, either the corresponding .img file's FAT tables aren't identical",
+                "or the tables_are_identical function doesn't work correctly"
+            )
+        );
+
+        // let's write the bee movie script to root.txt (why not), check, truncate the file, then check again
+        let mut file = fs.get_rw_file("root.txt").unwrap();
+
+        file.write_all(BEE_MOVIE_SCRIPT.as_bytes()).unwrap();
+        assert!(file.fs.FAT_tables_are_identical().unwrap());
+
+        file.seek(SeekFrom::Start(10_000)).unwrap();
+        assert!(file.fs.FAT_tables_are_identical().unwrap());
+    }
+
+    #[test]
+    #[expect(non_snake_case)]
+    fn FAT_tables_after_fat32_write_are_identical() {
+        use crate::fat::{BootRecord, Ebr};
+
+        use std::io::Cursor;
+
+        let mut storage = FromStd::new(Cursor::new(FAT32.to_owned())).unwrap();
+        let fs = FileSystem::new(&mut storage, FSOptions::new()).unwrap();
+
+        match &*fs.boot_record.borrow() {
+            BootRecord::Fat(boot_record_fat) => match &boot_record_fat.ebr {
+                Ebr::FAT32(ebr_fat32, _) => assert!(
+                    !ebr_fat32.extended_flags.mirroring_disabled(),
+                    "mirroring should be enabled for this .img file"
+                ),
+                _ => unreachable!(),
+            },
+            _ => unreachable!(),
+        }
+
+        assert!(
+            fs.FAT_tables_are_identical().unwrap(),
+            concat!(
+                "this should pass. ",
+                "if it doesn't, either the corresponding .img file's FAT tables aren't identical",
+                "or the tables_are_identical function doesn't work correctly"
+            )
+        );
+
+        // let's write the bee movie script to root.txt (why not), check, truncate the file, then check again
+        let mut file = fs.get_rw_file("hello.txt").unwrap();
+
+        file.write_all(BEE_MOVIE_SCRIPT.as_bytes()).unwrap();
+        assert!(file.fs.FAT_tables_are_identical().unwrap());
+
+        file.seek(SeekFrom::Start(10_000)).unwrap();
+        file.truncate().unwrap();
+        assert!(file.fs.FAT_tables_are_identical().unwrap());
+    }
+
+    #[test]
+    fn assert_img_fat_type() {
+        static TEST_CASES: &[(&[u8], FATType)] = &[
+            (MINFS, FATType::FAT12),
+            (FAT12, FATType::FAT12),
+            (FAT16, FATType::FAT16),
+            (FAT32, FATType::FAT32),
+        ];
+
+        for case in TEST_CASES {
+            use std::io::Cursor;
+
+            let mut storage = FromStd::new(Cursor::new(case.0)).unwrap();
+            let fs = FileSystem::new(&mut storage, FSOptions::new()).unwrap();
+
+            assert_eq!(fs.fat_type(), case.1)
+        }
+    }
+
+    #[test]
+    fn assert_fat_sector_size() {
+        static TEST_CASES: &[(&[u8], u16)] =
+            &[(MINFS, 512), (FAT12, 512), (FAT16, 512), (FAT32, 512)];
+
+        for case in TEST_CASES {
+            use std::io::Cursor;
+
+            let mut storage = FromStd::new(Cursor::new(case.0)).unwrap();
+            let sector_size = determine_fs_sector_size(&mut storage).unwrap();
+
+            assert_eq!(sector_size, case.1)
+        }
+    }
+
+    akin! {
+        let &fat_type = [FAT12, FAT16, FAT32];
+        let &unused_entries = [5, 2, 1];
+
+        #[test]
+        #[expect(non_snake_case)]
+        fn entry_defragment_~*fat_type() {
+            const UNUSED_ENTRY_COUNT: EntryCount = *unused_entries;
+
+            use std::io::Cursor;
+
+            let mut storage = FromStd::new(Cursor::new(~*fat_type.to_owned())).unwrap();
+            let fs = FileSystem::new(&mut storage, FSOptions::new()).unwrap();
+            fs.show_hidden(true);
+
+            // ik, this is dirty
+            let old_entry_count = {
+                let mut i: EntryCount = 0;
+
+                fs.go_to_dir("/").unwrap();
+
+                let mut current_entry = EntryLocation {
+                    unit: fs.dir_info.borrow().chain_start,
+                    index: 0,
+                };
+
+                while let Some(next_entry) = current_entry
+                    .next_entry(&fs)
+                    .unwrap()
+                    .filter(|entry| entry.entry_status(&fs).unwrap() != EntryStatus::LastUnused)
+                {
+                    current_entry = next_entry;
+                    i += 1
+                }
+
+                // we miss the last entry because of the .filter
+                i + 1
+            };
+
+            log::info!("Old entry count: {old_entry_count}");
+
+            let old_names: Box<[Box<str>]> = fs
+                .read_dir("/")
+                .unwrap()
+                .map(|entry| entry.unwrap())
+                .map(|entry| entry.path().file_name().unwrap().to_owned())
+                .map(Box::from)
+                .collect();
+
+            let new_entry_count = fs.defragment_entry_chain().unwrap();
+
+            log::info!("New entry count: {new_entry_count}");
+
+            let new_names: Box<[Box<str>]> = fs
+                .read_dir("/")
+                .unwrap()
+                .map(|entry| entry.unwrap())
+                .map(|entry| entry.path().file_name().unwrap().to_owned())
+                .map(Box::from)
+                .collect();
+
+            assert_eq!(old_names, new_names);
+            assert_eq!(old_entry_count - UNUSED_ENTRY_COUNT, new_entry_count);
+        }
+    }
+}
