+++ tests/common/I don't need a badge.txt
