+++ tests/common/imgs/minfs.img
